Scapy/requirements.txt
# Core dependencies
scapy>=2.5.0

# CLI and UI
typer>=0.9.0
rich>=13.0.0

# Configuration
pyyaml>=6.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Network utilities
netifaces>=0.11.0

# Optional: For better packet visualization
tabulate>=0.9.0
